Increase var size - 32bit instead of 16bit.
git-svn-id: http://forge.voodooprojects.org/svn/chameleon/trunk@1116 b5af796a-00a8-11df-919b-ffff7a100b7d

+++ i386/libsa/zalloc.c
@@ -46,7 +46,7 @@ typedef struct {
 
 static zmem * zalloced;
 static zmem * zavailable;
-static unsigned short  availableNodes, allocedNodes, totalNodes;
+static int  availableNodes, allocedNodes, totalNodes;
 static char * zalloc_base;
 static char * zalloc_end;
 static void   (*zerror)(char *, size_t, const char *, int);
